+++ core/src/test/java/cucumber/runtime/RuntimeTest.java
@@ -8,6 +8,7 @@ import gherkin.formatter.JSONPrettyFormatter;
 import gherkin.formatter.model.Step;
 import org.junit.Ignore;
 import org.junit.Test;
+import org.junit.internal.AssumptionViolatedException;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -133,6 +134,14 @@ public class RuntimeTest {
     public void non_strict_with_pending_steps() {
         Runtime runtime = createNonStrictRuntime();
         runtime.addError(new PendingException());
+
+        assertEquals(0x0, runtime.exitStatus());
+    }
+
+    @Test
+    public void non_strict_with_failed_junit_assumption() {
+        Runtime runtime = createNonStrictRuntime();
+        runtime.addError(new AssumptionViolatedException("should be treated like pending"));
 
         assertEquals(0x0, runtime.exitStatus());
     }